refactor(messaging): neutralize feishu dispatch and telegram helper naming
