+++ turbo/apps/api/src/signals/services/feishu-webhooks.service.ts
@@ -28,7 +28,7 @@ import {
   formatFeishuFileContext,
   type FeishuInboundMessage,
   type FeishuPromptFile,
-} from "./zero-feishu-dispatch.service";
+} from "./feishu-dispatch.service";
 import { publishFeishuOrgChanged } from "./feishu-realtime.service";
 
 const L = logger("FeishuWebhooks");
