+++ turbo/apps/api/src/signals/services/telegram-footer.service.ts
